organized logic tree modules
