+++ nzshm_model/logic_tree/__init__.py
@@ -1,0 +1,2 @@
+from .gmcm_logic_tree import GMCMBranch, GMCMBranchSet, GMCMLogicTree
+from .source_logic_tree import SourceBranch, SourceBranchSet, SourceLogicTree, SourceLogicTreeV1
