+++ nzshm_model/logic_tree/source_logic_tree/core/__init__.py
@@ -1,0 +1,1 @@
+from .fault_system_branch_set import BranchSetBase, BranchSetSpec
