+++ tests/test_slt_aliases.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 import pytest
 
-from nzshm_model.source_logic_tree.version2 import SourceLogicTree
+from nzshm_model.logic_tree import SourceLogicTree
 
 # import pytest
