+++ tests/test_slt_config.py
@@ -7,8 +7,8 @@ import dacite
 import pytest
 
 import nzshm_model
-from nzshm_model.source_logic_tree.version1 import Branch, FaultSystemLogicTree, SourceLogicTree
-from nzshm_model.source_logic_tree.version1.slt_config import (
+from nzshm_model.logic_tree.source_logic_tree.version1 import Branch, FaultSystemLogicTree, SourceLogicTree
+from nzshm_model.logic_tree.source_logic_tree.version1.slt_config import (
     decompose_crustal_tag,
     decompose_subduction_tag,
     from_config,
@@ -86,7 +86,7 @@ class TestStructure:
 class TestConfigSerialisation:
     @pytest.mark.parametrize("python_module", ['SLT_v9p0p0.py', 'SLT_v8_gmm_v2_final.py'])
     def test_slt_dict_to_json(self, python_module):
-        config = Path(__file__).parent.parent / 'nzshm_model' / 'source_logic_tree' / python_module
+        config = Path(__file__).parent.parent / 'nzshm_model' / 'logic_tree' / 'source_logic_tree' / python_module
         slt = from_config(config)
         obj = dataclasses.asdict(slt)
         jsonish = json.dumps(obj, indent=2)
@@ -96,7 +96,7 @@ class TestConfigSerialisation:
     @pytest.mark.parametrize("python_module", ['SLT_v9p0p0.py', 'SLT_v8_gmm_v2_final.py'])
     def test_slt_v8_round_trip(self, python_module):
 
-        config = Path(__file__).parent.parent / 'nzshm_model' / 'source_logic_tree' / python_module
+        config = Path(__file__).parent.parent / 'nzshm_model' / 'logic_tree' / 'source_logic_tree' / python_module
         slt = from_config(config)
         obj = dataclasses.asdict(slt)
         jsonish = json.dumps(obj, indent=2)
@@ -118,7 +118,9 @@ class TestConfigSerialisation:
 
 class TestFromConfig:
     def test_slt_v8(self):
-        config = Path(__file__).parent.parent / 'nzshm_model' / 'source_logic_tree' / 'SLT_v8_gmm_v2_final.py'
+        config = (
+            Path(__file__).parent.parent / 'nzshm_model' / 'logic_tree' / 'source_logic_tree' / 'SLT_v8_gmm_v2_final.py'
+        )
         slt = from_config(config)
         print(slt)
         assert slt.fault_system_lts[0].branches[-1].values[0].name == 'dm'
